fix: peer progress is an f64

+++ src/model/sync.rs
@@ -52,7 +52,7 @@ pub struct Peer {
     pub flags_desc: Option<String>,
     pub ip: Option<String>,
     pub port: Option<u16>,
-    pub progress: Option<u64>,
+    pub progress: Option<f64>,
     pub relevance: Option<u64>,
     pub up_speed: Option<u64>,
     pub uploaded: Option<u64>,
